Move pkg server_graph to network.server_graph
to allow for later addition of device graph and conversion / mapping
between them, turn prohibition algorithm, etc.
