+++ src/main/java/de/uni_kl/cs/discodnc/feedforward/analyses/PmooResults.java
@@ -33,8 +33,8 @@ package de.uni_kl.cs.discodnc.feedforward.analyses;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
 import de.uni_kl.cs.discodnc.feedforward.AnalysisResults;
+import de.uni_kl.cs.discodnc.network.server_graph.Server;
 import de.uni_kl.cs.discodnc.numbers.Num;
-import de.uni_kl.cs.discodnc.server_graph.Server;
 
 import java.util.Map;
 import java.util.Set;
