+++ src/main/java/de/uni_kl/cs/discodnc/feedforward/analyses/TandemMatchingResults.java
@@ -4,10 +4,10 @@ import java.util.Map;
 import java.util.Set;
 
 import de.uni_kl.cs.discodnc.numbers.Num;
-import de.uni_kl.cs.discodnc.server_graph.Server;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
 import de.uni_kl.cs.discodnc.feedforward.AnalysisResults;
+import de.uni_kl.cs.discodnc.network.server_graph.Server;
 
 public class TandemMatchingResults extends AnalysisResults {
 	protected Set<ServiceCurve> betas_e2e;
