+++ src/main/java/de/uni_kl/cs/discodnc/network/server_graph/ServerGraphFactory.java
@@ -26,7 +26,7 @@
  *
  */
 
-package de.uni_kl.cs.discodnc.server_graph;
+package de.uni_kl.cs.discodnc.network.server_graph;
 
 public interface ServerGraphFactory {
     ServerGraph createServerGraph();
